refactor(persistence): move database client to persistence layer
Centralize database client implementation in persistence layer for better organization and maintainability. The new location better reflects its purpose and improves code structure.
